build: add gulp build scripts
