+++ .eslintrc.js
@@ -76,5 +76,5 @@ module.exports = {
     'vue/require-default-prop': 'error',
     'vue/require-explicit-emits': 'error',
   },
-  ignorePatterns: ['dist', 'output', 'packages/site/src/router.ts', 'packages/site/src/sideNav.ts'],
+  ignorePatterns: ['dist', 'packages/site/src/router.ts', 'packages/site/src/sideNav.ts'],
 }
