+++ .ls-lint.yml
@@ -11,7 +11,12 @@ ls:
 
   scripts:
     .ts: camelCase
+    .config.ts: camelCase
     .svg: kebab-case
 
 ignore:
   - packages/components/i18n/src/locales
+  - packages/cdk/dist
+  - packages/cdk/node_modules
+  - packages/components/dist
+  - packages/components/node_modules
