--- gulpfile.ts
@@ -1,1 +1,0 @@
-import './scripts/gulp/gulpfile'
