+++ lerna.json
@@ -1,11 +1,12 @@
 {
   "packages": ["packages/*"],
-  "version": "0.0.0",
+  "version": "0.0.1",
   "command": {
     "publish": {
       "allowBranch": ["main", "next"],
       "ignoreChanges": ["ignored-file", "*.md"],
-      "message": "docs(release): publish %s"
+      "message": "docs(release): publish %s",
+      "registry": "https://registry.npmjs.org"
     },
     "bootstrap": {
       "ignore": "",
