+++ package.json
@@ -4,11 +4,13 @@
   "description": "A UI Component Library for Vue 3.x",
   "private": true,
   "scripts": {
-    "start": "gulp start",
-    "build": "gulp build",
-    "icons": "gulp icons",
-    "gen": "ts-node --project scripts/gen/tsconfig.json scripts/gen/generate.ts",
+    "start": "gulp start --gulpfile ./scripts/gulp/gulpfile.ts",
+    "build": "gulp build --gulpfile ./scripts/gulp/gulpfile.ts",
+    "icons": "gulp icons --gulpfile ./scripts/gulp/gulpfile.ts",
+    "gen": "ts-node scripts/gen/generate.ts",
     "postinstall": "lerna bootstrap",
+    "version": "lerna version --conventional-commits",
+    "publish": "lerna publish from-git",
     "test": "jest",
     "lint": "npm run ls-lint && npm run eslint && npm run stylelint && npm run markdownlint",
     "lint-fix": "npm run ls-lint && npm run eslint-fix && npm run stylelint-fix && npm run markdownlint-fix",
@@ -28,12 +30,15 @@
     "@commitlint/cli": "^12.0.0",
     "@commitlint/config-angular": "^12.0.0",
     "@ls-lint/ls-lint": "^1.9.0",
-    "@types/detect-port": "^1.3.0",
+    "@microsoft/api-extractor": "^7.13.0",
+    "@rollup/plugin-node-resolve": "^11.2.0",
+    "@rollup/plugin-typescript": "^8.2.0",
     "@types/figlet": "^1.2.1",
     "@types/fs-extra": "^9.0.0",
     "@types/gulp": "^4.0.0",
     "@types/inquirer": "^7.3.0",
     "@types/jest": "^26.0.0",
+    "@types/less": "^3.0.2",
     "@types/lodash": "^4.14.0",
     "@types/marked": "^2.0.0",
     "@types/parse5": "^6.0.0",
@@ -54,7 +59,8 @@
     "chalk": "^4.1.0",
     "codecov": "^3.8.0",
     "cz-conventional-changelog": "^3.3.0",
-    "detect-port": "^1.3.0",
+    "del": "^6.0.0",
+    "esbuild": "^0.10.0",
     "eslint": "^7.21.0",
     "eslint-config-prettier": "^8.1.0",
     "eslint-plugin-markdown": "^2.0.0",
@@ -63,13 +69,14 @@
     "figlet": "^1.5.0",
     "fs-extra": "^9.1.0",
     "gulp": "^4.0.0",
-    "gulp-clean": "^0.4.0",
     "husky": "^4.3.0",
     "inquirer": "^8.0.0",
     "jest": "^26.6.0",
     "jest-junit": "^12.0.0",
     "lerna": "^4.0.0",
     "less": "^4.1.0",
+    "less-plugin-clean-css": "^1.5.1",
+    "less-plugin-npm-import": "^2.1.0",
     "lint-staged": "^10.5.0",
     "lodash": "^4.17.0",
     "markdownlint-cli": "^0.27.0",
@@ -79,8 +86,9 @@
     "prettier": "^2.2.0",
     "prismjs": "^1.23.0",
     "remark": "^13.0.0",
-    "resolve-bin": "^0.4.0",
-    "rimraf": "^3.0.0",
+    "rollup-plugin-esbuild": "^3.0.0",
+    "rollup-plugin-typescript2": "^0.30.0",
+    "rollup-plugin-vue": "^6.0.0",
     "stylelint": "^13.12.0",
     "stylelint-config-prettier": "^8.0.0",
     "stylelint-config-standard": "^21.0.0",
