+++ packages/cdk/breakpoint/index.ts
@@ -1,1 +1,3 @@
-export * from './src/index'
+export * from './src/observe'
+export * from './src/breakpoints'
+export * from './src/mediaMatcher'
