--- packages/cdk/breakpoint/src/index.ts
@@ -1,3 +1,0 @@
-export * from './observe'
-export * from './breakpoints'
-export * from './mediaMatcher'
