+++ packages/cdk/forms/index.ts
@@ -2,6 +2,7 @@ export * from './src/controls/abstractControl'
 export * from './src/controls/formArray'
 export * from './src/controls/formControl'
 export * from './src/controls/formGroup'
+export * from './src/controls/types'
 export * from './src/types'
 export * from './src/useForms'
 export * from './src/useValueAccessor'
