+++ packages/cdk/subject/index.ts
@@ -1,2 +1,2 @@
-export { Subject } from './src/subject'
+export * from './src/subject'
 export * from './src/type'
