+++ packages/components/button/style/index.less
@@ -1,4 +1,5 @@
 @import '../../style/default.less';
+@import '../../icon/style/index.less';
 @import './mixin.less';
 
 @button-prefix: ~'@{idux-prefix}-button';
