+++ packages/components/empty/style/index.less
@@ -1,4 +1,5 @@
 @import '../../style/default.less';
+@import '../../icon/style/index.less';
 
 @empty-prefix: ~'@{idux-prefix}-empty';
