+++ packages/components/index.ts
@@ -104,49 +104,3 @@ export default {
   install,
   version,
 }
-
-export { useGlobalConfig } from './config'
-export { useLocale, addLocale, getLocale } from './i18n'
-export {
-  // export General
-  IxButton,
-  IxButtonGroup,
-  IxIcon,
-  IxTitle,
-  IxTypography,
-  // export Layout
-  IxDivider,
-  IxSpace,
-  IxRow,
-  IxCol,
-  // export Navigation
-  IxAffix,
-  // export Data Entry
-  IxCheckbox,
-  IxCheckboxGroup,
-  IxInput,
-  IxTextarea,
-  IxRadio,
-  IxRadioButton,
-  IxRadioGroup,
-  IxRate,
-  IxSwitch,
-  // export Data Display
-  IxBadge,
-  IxCard,
-  IxEmpty,
-  IxImage,
-  IxStatistic,
-  IxTimeline,
-  IxTimelineItem,
-  IxTooltip,
-  IxPopover,
-  // export Feedback
-  IxResult,
-  IxSpin,
-  IxProgress,
-  // export Other
-  IxBackTop,
-  // --- export end ---
-}
-export * from './message'
