+++ packages/components/package.json
@@ -5,9 +5,14 @@
   "author": "IduxFE<idux.feteam@gmail.com>",
   "license": "MIT",
   "scripts": {},
-  "main": "dist/index.js",
-  "module": "dist/index.es.js",
-  "unpkg": "dist/index.min.js",
+  "main": "./dist/index.umd.js",
+  "module": "./dist/index.es.js",
+  "exports": {
+    ".": {
+      "import": "./dist/index.es.js",
+      "require": "./dist/index.umd.js"
+    }
+  },
   "typings": "dist/index.d.ts",
   "style": "dist/theme/index.css",
   "keywords": [
@@ -32,6 +37,8 @@
     "vue": "^3.0.0"
   },
   "dependencies": {
-    "@idux/cdk": "^0.0.1"
+    "@idux/cdk": "^0.0.1",
+    "lodash": "^4.17.0",
+    "vue-types": "^3.0.0"
   }
 }
