+++ packages/components/result/style/index.less
@@ -1,5 +1,6 @@
 @import '../../style/default.less';
-@import "./mixins.less";
+@import '../../icon/style/index.less';
+@import './mixins.less';
 
 @result-prefix: ~'@{idux-prefix}-result';
 
@@ -29,7 +30,6 @@
 
   &-extra {
     margin-top: @result-extra-margin;
-
 
     & > * {
       margin-right: @result-extra-children-margin;
