+++ packages/site/vite.config.ts
@@ -5,25 +5,24 @@ import path from 'path'
 
 import { mdPlugin } from './plugins/mdPlugin'
 
-export default defineConfig({
-  plugins: [vuePlugin({ include: [/\.vue$/, /\.md$/] }), mdPlugin(), vueJsxPlugin()],
-  resolve: {
-    alias: [
-      {
-        find: '@idux/cdk',
-        replacement: path.resolve(__dirname, '../cdk'),
-      },
-      {
-        find: '@idux/components',
-        replacement: path.resolve(__dirname, '../components'),
-      },
-    ],
-  },
-  css: {
-    preprocessorOptions: {
-      less: {
-        javascriptEnabled: true,
+export default defineConfig(({ command }) => {
+  const isBuild = command === 'build'
+  const cdkResolve = isBuild ? '../cdk/dist' : '../cdk'
+  const componentsResolve = isBuild ? '../components/dist' : '../components'
+  return {
+    plugins: [vuePlugin({ include: [/\.vue$/, /\.md$/] }), mdPlugin(), vueJsxPlugin()],
+    resolve: {
+      alias: [
+        { find: '@idux/cdk', replacement: path.resolve(__dirname, cdkResolve) },
+        { find: '@idux/components', replacement: path.resolve(__dirname, componentsResolve) },
+      ],
+    },
+    css: {
+      preprocessorOptions: {
+        less: {
+          javascriptEnabled: true,
+        },
       },
     },
-  },
+  }
 })
