--- scripts/gen/tsconfig.json
@@ -1,6 +1,0 @@
-{
-  "compilerOptions": {
-    "lib": ["ESNext", "DOM"],
-    "types": ["node"]
-  }
-}
