+++ scripts/gulp/gulpConfig.ts
@@ -5,7 +5,7 @@ const packageRoot = join(__dirname, '../../packages')
 const componentsDirname = join(packageRoot, 'components')
 const siteDirname = join(packageRoot, 'site')
 
-export const buildConfig = {
+export const gulpConfig = {
   packageRoot,
   icon: {
     assetsDirname: join(__dirname, 'icons/assets'),
@@ -17,5 +17,10 @@ export const buildConfig = {
     docsDirname: join(siteDirname, 'src/docs'),
     sideNavFilename: join(siteDirname, 'src/sideNav.ts'),
     routerFilename: join(siteDirname, 'src/router.ts'),
+  },
+  build: {
+    cdkDirname: join(packageRoot, 'cdk'),
+    componentsDirname: join(packageRoot, 'components'),
+    declarationDirname: join(packageRoot, '../dist'),
   },
 } as const
