+++ scripts/gulp/icons/index.ts
@@ -1,53 +1,20 @@
-import { copy, existsSync, readdir, readFile, writeFile } from 'fs-extra'
-import { camelCase, upperFirst } from 'lodash'
-import { join } from 'path'
-import { optimize, OptimizeOptions } from 'svgo'
-import { buildConfig } from '../buildConfig'
+import { series, TaskFunction } from 'gulp'
+import { gulpConfig } from '../gulpConfig'
+import { clean } from '../taskHelpers'
+import { copyToSite, generateIcons } from './utils'
 
-const definitionTemplate = `export const {{definitionName}} = {
-  name: '{{name}}',
-  svgString: '{{svgString}}',
-}
-`
-const { assetsDirname, publicDirname, definitionsFilename } = buildConfig.icon
+const { publicDirname, definitionsFilename } = gulpConfig.icon
 
-const options: OptimizeOptions = {
-  plugins: [
-    { name: 'removeAttrs', params: { attrs: ['fill', 'class'] } },
-    { name: 'sortAttrs' },
-    { name: 'removeDimensions' },
-  ],
+const iconsClean = clean([publicDirname, definitionsFilename])
+
+const generate: TaskFunction = async done => {
+  await generateIcons()
+  done()
 }
 
-export async function generateIcons(): Promise<void> {
-  const iconPaths = await readdir(assetsDirname)
-  const definitionPromises = iconPaths.map(async iconName => {
-    const iconContent = await readFile(join(assetsDirname, iconName), 'utf8')
-    const { data } = optimize(iconContent, options)
+export const iconsGenerate = series(iconsClean, generate)
 
-    await writeFile(join(assetsDirname, iconName), data, 'utf8')
-
-    return getDefinition(iconName, data)
-  })
-
-  const definitions = await Promise.all(definitionPromises)
-
-  await writeFile(definitionsFilename, definitions.join('\n'), 'utf8')
-}
-
-function getDefinition(iconName: string, data: string) {
-  const _iconName = `${iconName.replace('.svg', '')}`
-  const camelCaseName = camelCase(_iconName)
-  const definitionName = upperFirst(camelCaseName)
-  return definitionTemplate
-    .replace('{{definitionName}}', definitionName)
-    .replace('{{name}}', _iconName)
-    .replace('{{svgString}}', data)
-}
-
-export async function copyToSite(): Promise<void> {
-  // 不存在的时候才 copy
-  if (!existsSync(publicDirname)) {
-    await copy(assetsDirname, publicDirname)
-  }
+export const iconsCopy: TaskFunction = async done => {
+  await copyToSite()
+  done()
 }
