--- scripts/gulp/tasks/site.ts
@@ -1,17 +1,0 @@
-import { task } from 'gulp'
-
-import { execNodeTask } from './taskHelpers'
-import { initSite } from '../site'
-
-task('site:init', done => {
-  initSite()
-  done()
-})
-
-task('site:serve', done => {
-  execNodeTask('lerna', 'lerna', ['run', 'start', '--stream'])(done)
-})
-
-task('site:build', done => {
-  execNodeTask('lerna', 'lerna', ['run', 'build', '--stream'])(done)
-})
