--- scripts/gulp/tsconfig.json
@@ -1,20 +1,0 @@
-{
-  "compilerOptions": {
-    "experimentalDecorators": true,
-    "noUnusedParameters": true,
-    "noUnusedLocals": true,
-    "lib": ["es2015", "dom"],
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "strictNullChecks": true,
-    "esModuleInterop": true,
-    "strictFunctionTypes": true,
-    "noImplicitThis": true,
-    "noEmitOnError": true,
-    "noImplicitAny": true,
-    "target": "es6",
-    "types": ["node"],
-    "baseUrl": "."
-  },
-  "files": ["gulpfile.ts"]
-}
